requirements.txt
bokeh==0.13.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
DateTime==4.2
dill==0.2.8.2
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.1
packaging==17.1
pandas==0.23.4
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
simplejson==3.16.0
six==1.11.0
tornado==5.1
urllib3==1.23
Werkzeug==0.14.1
zope.interface==4.5.0
